# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SmartDNS Control Plane API",
  "version": "1.0.0",
  "description": "Публичный API панели управления. Веб-интерфейс использует ровно эти же эндпоинты — приватного канала нет.\n\n**Аутентификация.** Сессионная cookie `sdns_session` (HttpOnly, SameSite=Lax) плюс заголовок `X-CSRF-Token` на всех изменяющих методах. Токен возвращают `POST /auth/login` и `GET /auth/me`.\n\n**Одновременные изменения.** Изменяемые объекты содержат поле `version`. Передайте его в `If-Match`, чтобы получить 412 вместо перезаписи чужих изменений.\n\n**Повторные попытки.** `Idempotency-Key` на регистрации нод, сборке, выкате, откате и ручном обновлении списков: повтор возвращает сохранённый ответ, а не выполняет действие второй раз.\n\n**Время** — RFC 3339 в UTC. **Постраничность** — курсорная.\n\nAPI агентов (`/agent/v1/…`) обслуживается отдельным слушателем со взаимным TLS и здесь не описан: он предназначен только для нод. Его контракт — в `docs/API.md`.",
  "license": {
   "name": "См. LICENSE"
  }
 },
 "servers": [
  {
   "url": "/api/v1",
   "description": "Тот же origin, что и веб-панель"
  }
 ],
 "security": [
  {
   "sessionCookie": []
  }
 ],
...
 "paths": {
...
  "/auth/sessions": {
   "get": {
    "tags": [
     "Аутентификация"
    ],
    "summary": "Активные сессии",
    "description": "Список сессий пользователя.",
    "responses": {
     "200": {
      "description": "Успех"
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "401": {
      "description": "Не аутентифицирован",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав или не пройдена CSRF-проверка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт: объект используется или изменён другим клиентом",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "412": {
      "description": "Версия объекта устарела (If-Match)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "422": {
      "description": "Конфигурация не компилируется",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   }
  },
  "/auth/sessions/{id}": {
   "delete": {
    "tags": [
     "Аутентификация"
    ],
    "summary": "Завершить сессию",
    "description": "Отзывает одну сессию.",
    "responses": {
     "200": {
      "description": "Успех"
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "401": {
      "description": "Не аутентифицирован",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав или не пройдена CSRF-проверка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт: объект используется или изменён другим клиентом",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "412": {
      "description": "Версия объекта устарела (If-Match)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "422": {
      "description": "Конфигурация не компилируется",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ]
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "sessionCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "sdns_session"
   }
  },
  "schemas": {
   "Error": {
    "type": "object",
    "required": [
     "code",
     "message",
     "request_id"
    ],
    "properties": {
     "code": {
      "type": "string",
      "examples": [
       "invalid_request"
      ]
     },
     "message": {
      "type": "string"
     },
     "details": {
      "description": "Структурированный контекст: список зависимостей при 409 или список конфликтов правил."
     },
     "request_id": {
      "type": "string"
     }
    }
   },
...
  }
 }
}